refactor(utils): Move socket.js file to /utils folder
