+++ controllers/feed.js
@@ -3,7 +3,7 @@ const path = require('path');
 
 const { validationResult } = require('express-validator/check');
 
-const io = require('../socket');
+const io = require('../utils/socket');
 const Post = require('../models/post');
 const User = require('../models/user');
