+++ utils/socket.js
@@ -16,14 +16,4 @@ const getIO = () => {
 module.exports = {
   init,
   getIO
-  // init(httpServer) {
-  //   io = require('socket.io')(httpServer);
-  //   return io;
-  // }
-  // getIO() {
-  //   if (!io) {
-  //     throw new Error('Socket.io not initialized');
-  //   }
-  //   return io;
-  // }
 };
